requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.5
chardet==4.0.0
cssselect2==0.4.1
defusedxml==0.7.1
discord-py-slash-command==1.1.0
discord.py==1.6.0
emojis==0.6.0
idna==3.1
multidict==5.1.0
Pillow==8.1.2
pycparser==2.20
tinycss2==1.1.0
typing-extensions==3.7.4.3
webencodings==0.5.1
yarl==1.6.3
